Pair styles to components
