+++ client/src/components/Button.scss
@@ -1,5 +1,35 @@
 @import './Colors.scss';
 
+button,
+.aButton {
+  padding: 10px 20px;
+  background-color: $darkBlue;
+  border: 1px solid $bodyGreen;
+  font-size: 16px;
+  font-weight: 500;
+  color: $bodyGreen;
+  margin-top: 1rem;
+}
+
+button:hover {
+  cursor: pointer;
+}
+
+button:focus {
+  outline: none;
+}
+
+form button {
+  padding: 12px 20px;
+  margin: 24px;
+  background-color: $darkBlue;
+  border: 1px solid $bodyGreen;
+  font-size: 16px;
+  font-weight: 500;
+  color: $bodyGreen;
+}
+
+
 .labels-container {
     display: flex;
     flex-direction: row-reverse;
