+++ client/src/components/HomePage.scss
@@ -1,0 +1,30 @@
+@import './Colors.scss';
+
+main {
+    max-width: 1140px;
+    margin: 0 auto;
+    min-height: 72vh;
+  }
+  
+  h1,
+  h2 {
+    margin: 0 auto;
+  }
+  
+  h1 {
+    font-weight: 500;
+    font-size: 24px;
+  }
+  
+  h2 {
+    font-weight: 300;
+    font-size: 18px;
+    margin-top: 16px;
+  }
+  
+  @media (max-width: 600px) {
+    main {
+      margin-left: .5rem;
+      margin-right: .5rem;
+    }
+  }
