+++ client/src/components/Navbar.scss
@@ -1,0 +1,56 @@
+@import './Colors.scss';
+
+#root {
+    padding-top: 104px;
+    margin: 0 auto;
+  }
+  
+nav {
+width: 100%;
+padding: 16px 40px;
+background-color: $darkBlue;
+position: fixed;
+top: 0;
+display: flex;
+flex-direction: row;
+align-items: center;
+justify-content: space-between;
+z-index: 999;
+}
+
+nav div#logoHome img {
+height: 28px;
+}
+
+nav div:nth-child(2) a:not(:last-child) {
+margin-right: 24px;
+}
+
+.loggedUser .nav-bar-texts{
+display: inline-block;
+margin-right: 24px;
+}
+
+.loggedUser p {
+display: inline-block;
+margin-right: 24px;
+font-weight: 300;
+}
+
+.loggedUser span {
+color:$bodyGreen;
+}
+
+/* hide burger btn if larger than 601px */
+@media (min-width: 601px) {
+.navbar-toggle-btn {
+    display: none;
+}
+}
+
+/* hide loggedUser navbar items if below 600px */
+@media (max-width: 600px) {
+.loggedUser {
+    display: none;
+}
+}
