+++ client/src/components/Search.scss
@@ -1,0 +1,40 @@
+@import './Colors.scss';
+
+form#search {
+    margin-top: 48px;
+    margin-bottom: 80px;
+  }
+  
+  #search input {
+    padding: 16px;
+    border: 0;
+    font-size: 16px;
+    font-weight: 500;
+    width: 60%;
+    max-width: 600px;
+    min-width: 400px;
+    background-color: $darkerBlue;
+    color: $almostWhite;
+  }
+  
+  #search input::placeholder {
+    color: $gray;
+    font-weight: 300;
+  }
+  
+  #search input:focus {
+    outline-color: $bodyGreen;
+    outline-width: 2px;
+    color:$almostWhite;
+  }
+
+  @media all and (max-width: 600px) {
+    form#search {
+      margin-top: 30px;
+      margin-bottom: 30px;
+    }
+    #search input {
+      width: 90%;
+      min-width: auto;
+    }
+  }
