+++ client/src/components/WishList.js
@@ -1,49 +1,47 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { fetchAllApps } from '../services/app';
 import appIconPlaceholder from '../app-icon-placeholder.svg';
+import Loader from './Loader';
 
-// fetch all apps, and filter to wishUser_id includes props user id
-// TODO: conver to class
-// TODO: remove from wish list
-class WishList extends Component {
-    state = {
-      appList: [],
-    };
+// fetch all apps, filter to wishUser_id includes props user id
+function WishListHook(props) {
+  const [appList, setAppList] = useState([]);
 
-    componentDidMount() {
-      fetchAllApps()
-        .then((apps) => {
-          this.setState({
-            appList: apps,
-          });
-        })
-        .catch((error) => {
-          alert(error.message);
-        });
-    }
+  useEffect(() => {
+    fetchAllApps()
+      .then((apps) => {
+        setAppList(apps);
+      })
+      .catch((error) => {
+        alert(error.message);
+      });
+  }, [setAppList]);
 
-    render() {
-      const filteredApps = this.state.appList
-        .filter((app) => app.wishUser.includes(this.props.user._id))
-        .map((app) => (
-          <div key={app._id} className="appCard_wish">
-            <Link to={`/apps/${app._id}`}>
-              <img src={app.logo || appIconPlaceholder} alt="" style={{ width: '50px' }} />
-              <h3>{app.name}</h3>
-              <h6>{app.category.name}</h6>
-            </Link>
-          </div>
-        ));
-      return (
-        <div>
-          <h1>My wish list</h1>
-          <section id="listContainer">
-            {filteredApps}
-          </section>
-        </div>
-      );
-    }
+  if (!appList) return <Loader />;
+
+  return (
+    <div>
+      <h1>My wish list</h1>
+      <section id="listContainer">
+        {appList
+          .filter((app) => app.wishUser.includes(props.user._id))
+          .map((app) => (
+            <div key={app._id} className="appCard">
+              <Link to={`/apps/${app._id}`}>
+                <img src={app.logo || appIconPlaceholder} alt="" />
+              </Link>
+              <div>
+                <Link to={`/apps/${app._id}`}>
+                  <h3>{app.name}</h3>
+                </Link>
+                <h6>{app.category.name}</h6>
+              </div>
+            </div>
+          ))}
+      </section>
+    </div>
+  );
 }
 
-export default WishList;
+export default WishListHook;
