+++ client/src/components/currentUnused/AlternativesPage.js
