+++ client/src/components/style.scss
@@ -6,8 +6,6 @@
 
 /* IMPORTS
 =====================*/
-@import "~react-alice-carousel/lib/alice-carousel.css";
-
 @import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;500;700&display=swap");
 
 @import './Colors.scss';
@@ -26,7 +24,6 @@ html {
   /*   border: 1px solid red;
  */
 }
-
 
 body {
   margin: 0 auto;
@@ -50,297 +47,3 @@ a:hover {
   color:$bodyGreen;
 }
 
-button,
-.aButton {
-  padding: 10px 20px;
-  background-color: $darkBlue;
-  border: 1px solid $bodyGreen;
-  font-size: 16px;
-  font-weight: 500;
-  color: $bodyGreen;
-  margin-top: 1rem;
-}
-
-button:hover {
-  cursor: pointer;
-}
-
-button:focus {
-  outline: none;
-}
-
-form button {
-  padding: 12px 20px;
-  margin: 24px;
-  background-color: $darkBlue;
-  border: 1px solid $bodyGreen;
-  font-size: 16px;
-  font-weight: 500;
-  color: $bodyGreen;
-}
-
-/* HEADER
-=====================*/
-
-#root {
-  padding-top: 104px;
-  margin: 0 auto;
-}
-
-nav {
-  width: 100%;
-  padding: 16px 40px;
-  background-color: $darkBlue;
-  position: fixed;
-  top: 0;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  z-index: 999;
-}
-
-nav div#logoHome img {
-  height: 28px;
-}
-
-nav div:nth-child(2) a:not(:last-child) {
-  margin-right: 24px;
-}
-
-.loggedUser .nav-bar-texts{
-  display: inline-block;
-  margin-right: 24px;
-}
-
-.loggedUser p {
-  display: inline-block;
-  margin-right: 24px;
-  font-weight: 300;
-}
-
-.loggedUser span {
-  color:$bodyGreen;
-}
-
-/* hide burger btn if larger than 601px */
-@media (min-width: 601px) {
-  .navbar-toggle-btn {
-    display: none;
-  }
-}
-
-/* hide loggedUser navbar items if below 600px */
-@media (max-width: 600px) {
-  .loggedUser {
-    display: none;
-  }
-}
-
-/* MAIN
-=====================*/
-
-main {
-  max-width: 1140px;
-  margin: 0 auto;
-  min-height: 72vh;
-}
-
-h1,
-h2 {
-  margin: 0 auto;
-}
-
-h1 {
-  font-weight: 500;
-  font-size: 24px;
-}
-
-h2 {
-  font-weight: 300;
-  font-size: 18px;
-  margin-top: 16px;
-}
-
-@media (max-width: 600px) {
-  main {
-    margin-left: .5rem;
-    margin-right: .5rem;
-  }
-}
-
-/* SEARCH */
-
-form#search {
-  margin-top: 48px;
-  margin-bottom: 80px;
-}
-
-#search input {
-  padding: 16px;
-  border: 0;
-  font-size: 16px;
-  font-weight: 500;
-  width: 60%;
-  max-width: 600px;
-  min-width: 400px;
-  background-color: $darkerBlue;
-  color: $almostWhite;
-}
-
-#search input::placeholder {
-  color: $gray;
-  font-weight: 300;
-}
-
-#search input:focus {
-  outline-color: $bodyGreen;
-  outline-width: 2px;
-  color:$almostWhite;
-}
-
-/* CATEGORIES */
-
-#catContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
-
-button.btnCategories {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 112px;
-  background: none;
-  border: none;
-  padding: 0;
-  font-size: 14px;
-  font-weight: 300;
-}
-
-button.btnCategories:focus {
-  outline: none;
-}
-
-button.btnCategories img {
-  height: 40px;
-  margin-bottom: 14px;
-}
-
-button.btnCategories p {
-  text-align: center;
-  font-size: 14px;
-  color:$bodyGreen;
-}
-
-/* LIST */
-
-#listContainer {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin: 88px 0;
-}
-
-.appCard {
-  padding: 16px 16px;
-  margin: 8px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 30%;
-  text-align: left;
-  background-color: $darkBlue;
-  box-shadow: 0px 1px 4px $boxShadow;
-}
-
-.appCard a {
-  width: 100px;
-}
-
-.appCard a img {
-  height: 64px;
-  display: block;
-  margin: 0 auto;
-  max-width: 88px;
-  filter: saturate(0.8);
-}
-
-.appCard div {
-  margin-left: 16px;
-}
-
-.appCard div a {
-  display: inline-block;
-  
-}
-.appCard h3 {
-  font-size: 19px;
-  margin-bottom: 8px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.appCard h6 {
-  color:$lightGray;
-  font-size: 12px;
-}
-
-.fadeIn {
-  animation-name: fadeIn;
-  animation-duration: .3s; 
-  animation-iteration-count: 1;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-@media all and (max-width: 600px) {
-  .appCard {
-    width: 80%;
-  }
-  .appCard a {
-    width: 136px;
-  }
-  .appCard h3 {
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    overflow: hidden;
-  }
-  #listContainer {
-    margin: 10px 0 50px 0;
-  }
-  form#search {
-    margin-top: 30px;
-    margin-bottom: 48px;
-  }
-  #search input {
-    width: 90%;
-    min-width: auto;
-  }
-}
-
-/* FOOTER
-=====================*/
-
-/* TODO: sticky footer */
-footer {
-  font-weight: 300;
-  font-size: 14px;
-  margin: 24px;
-}
-
-footer p:first-child {
-  margin-bottom: 8px;
-}
-footer a {
-  color: $bodyGreen;
-}
