+++ client/src/index.js
@@ -8,6 +8,12 @@ import './components/AppDetail.scss';
 import './components/NewApp.scss';
 import './components/Medias.scss';
 import './components/Button.scss';
+import './components/Categories.scss';
+import './components/List.scss';
+import './components/Navbar.scss';
+import './components/Search.scss';
+import './components/Footer.scss';
+import './components/HomePage.scss';
 
 import App from './App';
 import * as serviceWorker from './serviceWorker';
